Create Sites with planning permissions
Task Force is focused to address supply-related issues to the delivery of units of houses in the Dublin area. The group are working together (Goverment and Dublin 4 local Planning authorities) to identify housing developments, which has approved planning permission and that are capable of delivering housing to the market.

Through this study is intended to be able to predict which houses(units) with planning permissions, in terms of the activity on site, which ones can be avaible to meet the market demand of houses in future, taking in consideration a dataset from Q1 2018.